strspn

+++ 0x07-pointers_arrays_strings/3-strspn.c
@@ -1,5 +1,6 @@
 #include "main.h"
 #include <string.h>
+#include <stdio.h>
 
 /**
  * _strspn - gets the length of a prefix substring
@@ -10,7 +11,7 @@
 
 unsigned int _strspn(char *s, char *accept)
 {
-	unsigned int *len;
+	long unsigned int *len;
 
 	len = strspn(s, accept);
